- Fixed warning in SoxrAdapter - Updated include statement in LinearResampler

+++ mt32emu_qt/src/resample/SoxrAdapter.cpp
@@ -99,6 +99,6 @@ void SoxrAdapter::getOutputSamples(Sample *buffer, uint length) {
 			qDebug() << "SampleRateConverter: got 0 frames from SOXR, weird";
 		}
 		buffer += gotFrames << 1;
-		length -= gotFrames;
+		length -= (uint)gotFrames;
 	}
 }
